+++ global.json
@@ -1,6 +1,6 @@
 {
   "sdk": {
-    "version": "6.0.100-preview.5.21302.13",
+    "version": "6.0.100-preview.6.21324.1",
     "allowPrerelease": false,
     "rollForward": "latestMajor"
   }
